+++ packages/guide/src/app/shared/components/item-icon/item-icon.component.ts
@@ -1,6 +1,7 @@
 import { Component, HostBinding, inject, Input, numberAttribute, ViewEncapsulation } from '@angular/core';
 import { Quality } from '@ci/data-types';
 import { SettingsService } from '../../services/settings.service';
+import { GAME_VERSION } from "../../../core/injection-tokens/version.injection-token";
 
 @Component({
     selector: 'app-item-icon',
@@ -13,11 +14,8 @@ export class ItemIconComponent {
     @Input() subIconName?: string | null;
     @Input() quality?: Quality | undefined;
     @Input({transform: numberAttribute}) amount = 0;
-    protected version: string;
-    private readonly _settings = inject(SettingsService);
+    protected environment = inject(SettingsService).getSettings().useBeta ? 'beta' : 'live';
+    protected version = inject(GAME_VERSION);
     @HostBinding('class.app-item-icon') private _setClass = true;
 
-    constructor() {
-        this.version = this._settings.getSettings().useBeta ? 'beta' : 'live';
-    }
 }