+++ packages/guide/src/app/shared/components/ui-icon/ui-icon.component.ts
@@ -1,5 +1,6 @@
-import { Component, HostBinding, input, ViewEncapsulation } from '@angular/core';
+import { Component, HostBinding, inject, input, ViewEncapsulation } from '@angular/core';
 import { UiIcon } from '@ci/data-types';
+import { GAME_VERSION } from "../../../core/injection-tokens/version.injection-token";
 
 @Component({
     selector: 'app-ui-icon',
@@ -11,6 +12,7 @@ import { UiIcon } from '@ci/data-types';
 })
 export class UiIconComponent {
 
+    protected version = inject(GAME_VERSION);
     uiIcon = input.required<UiIcon>();
     @HostBinding('class.app-ui-icon') private _setClass = true;