+++ packages/guide/src/app/shared/components/database-item-details/database-item-details.component.ts
@@ -7,6 +7,7 @@ import {
     input,
     signal,
     TemplateRef,
+    untracked,
     WritableSignal
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
@@ -22,6 +23,7 @@ import { ToDoToggleComponent } from "../to-do-toggle/to-do-toggle.component";
 import { ToDoContext } from "../../../core/types/to-do-context.type";
 import { ListDetailService } from "../list-detail-container/list-detail.service";
 import { UiIconComponent } from "../ui-icon/ui-icon.component";
+import { MatProgressSpinner } from "@angular/material/progress-spinner";
 
 
 @Component({
@@ -36,7 +38,8 @@ import { UiIconComponent } from "../ui-icon/ui-icon.component";
         DatabaseItemDetailsDirective,
         SharedModule,
         ToDoToggleComponent,
-        UiIconComponent
+        UiIconComponent,
+        MatProgressSpinner
     ],
 
 })
@@ -48,21 +51,33 @@ export class DatabaseItemDetailsComponent {
     amount = input<number>();
     quality = input<Quality>();
     @ContentChild(TemplateRef) databaseItemDetails: TemplateRef<any> | null = null;
-    protected databaseItem: WritableSignal<DatabaseItem | undefined> = signal(undefined)
+    protected databaseItem: WritableSignal<DatabaseItem | undefined> = signal(undefined);
+    protected isFetching = signal(false)
     protected readonly UiIcon = UiIcon;
     protected readonly uiIcon = UiIcon;
     protected readonly listDetails = inject(ListDetailService);
     private readonly database = inject(DatabaseService);
 
+    #timer: ReturnType<typeof setTimeout> | undefined;
+
     constructor() {
         effect(() => {
+            this.#timer = setTimeout(() => untracked(() => this.isFetching.set(true)), 150)
             this.database
                 .fetchDatabaseItem$(this.itemId())
                 .pipe(
-                    take(1)
+                    take(1),
                 ).subscribe({
                 next: i => {
-                    this.databaseItem.set(i)
+                    untracked(() => {
+                        clearTimeout(this.#timer)
+                        this.isFetching.set(false)
+                        this.databaseItem.set(i);
+                    })
+                },
+                error: () => {
+                    clearTimeout(this.#timer)
+                    untracked(() => this.isFetching.set(false))
                 }
             })
         }, {allowSignalWrites: true});