+++ packages/data-types/src/index.ts
@@ -113,5 +113,6 @@ export * from './lib/types/treasure-hunt.type';
 export * from './lib/types/weather.type';
 export * from './lib/types/item-mixing-recipe-data.type';
 
+export * from './lib/dashboard/critter-dashboard-entry.type';
 export * from './lib/dashboard/fish-dashboard-entry.type';
 export * from './lib/dashboard/birthday-dashboard-entry.type';