+++ packages/data-types/src/lib/dashboard/critter-dashboard-entry.type.ts
@@ -1,0 +1,8 @@
+import { Season, Weather } from "@ci/data-types";
+
+export type CritterDashboardEntry = {
+    id: string,
+    iconName: string | null,
+    seasons: Season[],
+    weathers: Weather[]
+}