+++ packages/guide/src/app/my-coral-guide/services/dashboard.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, signal } from '@angular/core';
 import { BaseDbService } from "../../shared/services/base-db.service";
 import { Observable, of, shareReplay, tap } from "rxjs";
-import { BirthdayDashboardEntry, FishDashboardEntry } from "@ci/data-types";
+import { BirthdayDashboardEntry, CritterDashboardEntry, FishDashboardEntry } from "@ci/data-types";
 
 @Injectable({
     providedIn: 'root'
@@ -9,6 +9,8 @@ import { BirthdayDashboardEntry, FishDashboardEntry } from "@ci/data-types";
 export class DashboardService extends BaseDbService {
 
     #fish = signal<FishDashboardEntry[]>([]);
+    #insects = signal<CritterDashboardEntry[]>([]);
+    #oceanCritter = signal<CritterDashboardEntry[]>([]);
     #birthdays = signal<BirthdayDashboardEntry[]>([]);
 
     getFish$(): Observable<FishDashboardEntry[]> {
@@ -23,8 +25,40 @@ export class DashboardService extends BaseDbService {
         }
     }
 
+    getInsects$(): Observable<CritterDashboardEntry[]> {
+        if (!this.#insects().length) {
+            return this.http.get<CritterDashboardEntry[]>(`${this.BASE_PATH}/dashboards/museum-insects.json`)
+                .pipe(
+                    tap(items => this.#insects.set(items)),
+                    shareReplay(1)
+                );
+        } else {
+            return of((this.#insects.asReadonly()()))
+        }
+    }
+
+    getOceanCritters$(): Observable<CritterDashboardEntry[]> {
+        if (!this.#oceanCritter().length) {
+            return this.http.get<CritterDashboardEntry[]>(`${this.BASE_PATH}/dashboards/museum-ocean-critters.json`)
+                .pipe(
+                    tap(items => this.#oceanCritter.set(items)),
+                    shareReplay(1)
+                );
+        } else {
+            return of((this.#oceanCritter.asReadonly()()))
+        }
+    }
+
     getFish() {
         return this.#fish.asReadonly();
+    }
+
+    getInsects() {
+        return this.#insects.asReadonly();
+    }
+
+    getOceanCritter() {
+        return this.#oceanCritter.asReadonly();
     }
 
     getBirthdays$(): Observable<BirthdayDashboardEntry[]> {